requirements.txt
flask
flask-cors
numpy
dotenv
psycopg2-binary
#tensorflow==2.17.0
gunicorn
joblib
pandas
seaborn
matplotlib
imbalanced-learn
resend